Use ShadowEnemyTactic in DefensePlay

+++ src/software/ai/hl/stp/play/defense/defense_play_base.h
@@ -36,14 +36,14 @@ class DefensePlayFSMBase
      *
      * @param num_crease_defenders the number of crease defender tactics to set
      */
-    void setUpCreaseDefenders(int num_crease_defenders);
+    void setUpCreaseDefenders(size_t num_crease_defenders);
 
     /**
      * Helper function to set up pass defender tactic vector members
      *
      * @param num_pass_defenders the number of pass defender tactics to set
      */
-    void setUpPassDefenders(int num_pass_defenders);
+    void setUpPassDefenders(size_t num_pass_defenders);
 
     /**
      * Helper function to set up alignments for the crease defense robots and control
